+++ src/Commands/1_music/stop.js
@@ -2,18 +2,18 @@ const Command = require('../../Structures/Command');
 
 const { Permissions } = require('discord.js');
 const queue = require('../../Data/queue');
-const { emoji: { success, warning }, response: { wrongChannel, noMusic } } = require('../../../config/config.json');
+const { bot: { ownerID }, emoji: { success, warning }, response: { wrongChannel, noMusic } } = require('../../../config/config.json');
 
 
 module.exports = new Command({
 	name: 'stop',
 	aliases: [ 'dc', 'disconnect', 'fuckoff' ],
-	description: 'Stops the current song and clears the queue.',
+	description: 'Stops the current song, clears the queue and disconnects the client.',
 	async run(message, args, client) {
 		const guildQueue = queue.get(message.guild.id);
         if (!guildQueue) return message.channel.send(`${warning} ${noMusic}`);
 
-        if (( !message.member.voice.channel || guildQueue.connection.joinConfig.channelId != message.member.voice.channel.id ) && !message.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) return message.channel.send(`${warning} ${wrongChannel}`);
+        if (( !message.member.voice.channel || guildQueue.connection.joinConfig.channelId != message.member.voice.channel.id ) && !message.member.permissions.has(Permissions.FLAGS.MANAGE_CHANNELS) && message.author.id != ownerID) return message.channel.send(`${warning} ${wrongChannel}`);
 
 
 		guildQueue.updater.emitter.emit('end');